requirements.txt
asgiref==3.5.2
attrs==22.1.0
autopep8==2.0.0
certifi==2022.12.7
charset-normalizer==2.1.1
Django==4.1.4
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
exceptiongroup==1.0.4
factory-boy==3.2.1
Faker==15.3.4
idna==3.4
inflection==0.5.1
iniconfig==1.1.1
packaging==22.0
pluggy==1.0.0
pycodestyle==2.10.0
PyJWT==2.6.0
pytest==7.2.0
pytest-django==4.5.2
pytest-factoryboy==2.5.1
python-dateutil==2.8.2
python-decouple==3.6
pytz==2022.6
requests==2.28.1
six==1.16.0
sqlparse==0.4.3
tomli==2.0.1
typing_extensions==4.4.0
urllib3==1.26.13
